requirements.txt
python-telegram-bot==21.6
python-dotenv==1.0.1
pydantic==2.6.4
pydantic-settings==2.2.1
pydantic_core==2.16.3
loguru==0.7.2